+++ src/client/impl/SceneImpl.h
@@ -190,7 +190,7 @@ namespace ramses::internal
         bool createTextureConsumer(const TextureSamplerMS& sampler, dataConsumerId_t id);
         bool createTextureConsumer(const TextureSamplerExternal& sampler, dataConsumerId_t id);
 
-        ArrayBuffer* createArrayBuffer(ramses::EDataType dataType, uint32_t maxNumElements, std::string_view name);
+        ArrayBuffer* createArrayBuffer(ramses::EDataType dataType, size_t maxNumElements, std::string_view name);
         Texture2DBuffer* createTexture2DBuffer (size_t mipLevels, uint32_t width, uint32_t height, ETextureFormat textureFormat, std::string_view name);
 
         SceneReference* createSceneReference(sceneId_t referencedScene, std::string_view name);
@@ -208,7 +208,7 @@ namespace ramses::internal
 
         template <typename T>
         // NOLINTNEXTLINE(modernize-avoid-c-arrays)
-        ramses::ArrayResource* createArrayResource(uint32_t numElements, const T* arrayData, std::string_view name);
+        ramses::ArrayResource* createArrayResource(size_t numElements, const T* arrayData, std::string_view name);
         Texture2D* createTexture2D(uint32_t width, uint32_t height, ETextureFormat format, const std::vector<MipLevelData>& mipLevelData, bool generateMipChain, const TextureSwizzle& swizzle, std::string_view name);
         Texture3D* createTexture3D(uint32_t width, uint32_t height, uint32_t depth, ETextureFormat format, const std::vector<MipLevelData>& mipLevelData, bool generateMipChain, std::string_view name);
         TextureCube* createTextureCube(uint32_t size, ETextureFormat format, const std::vector<CubeMipLevelData>& mipLevelData, bool generateMipChain, const TextureSwizzle& swizzle, std::string_view name);
